pyproject.toml
[tool.poetry]
name = "gift-autobuyer"
version = "0.1.0"
description = ""
authors = ["kirillketrik <kirillketrik34@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.40.0"
coloredlogs = "^15.0.1"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
pyyaml = "^6.0.2"
aiohttp = "^3.12.13"
tortoise-orm = "^0.25.1"
g4f = "^0.5.6.3"
loguru = "^0.7.3"
aiogram = "^3.21.0"
aiogram-dialog = "^2.4.0"
apscheduler = "^3.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
